+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Added Helpers\Display::attributes and Helpers\Display::data_attributes - render attributes for HTML elements
 - Added Acf\LocalJsonByLocationParam - sync local JSON for groups that match a specific location rule
+- Added Acf\LocalJsonCustomTemplate - sync local JSON for a specific custom page template
 
 ## [0.4.0] - 2023-06-28